# docs/openapi.json
{
  "components": {
    "schemas": {
...
      "ClaimRunRequest": {
        "properties": {
          "labeler": {
            "title": "Labeler",
            "type": "string"
          }
        },
        "required": [
          "labeler"
        ],
        "title": "ClaimRunRequest",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ScamTypeCatalogRequest": {
        "properties": {
          "description": {
            "anyOf": [
              {
                "maxLength": 200,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "labels": {
            "items": {
              "type": "string"
            },
            "title": "Labels",
            "type": "array"
          },
          "name": {
            "maxLength": 80,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "name"
        ],
        "title": "ScamTypeCatalogRequest",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "**한국어 사기 신호 검출 reference implementation** — 음성·텍스트·이미지·PDF·URL 입력에서 위험 신호를 검출하고 각 신호의 학술/법적 근거를 transparent 하게 보고한다.\n\n⚠️ **Identity Boundary** (CLAUDE.md): ScamGuardian 은 *사기 판정을 내리지 않는다*. VirusTotal 이 70개 백신의 검출 결과를 보고만 하는 모델과 동일. 판정 logic 은 통합한 기업(통신사·은행·메신저 앱)이 자기 risk tolerance 에 따라 구현한다.\n\n- **Public 4 endpoint** (`Public` 태그) — 외부 통합 시작점, `DetectionReport` 응답\n- **Webhook** (`Webhook` 태그) — 카카오 오픈빌더\n- **Admin** (3 태그) — 라벨링 / platform / training\n- **Live Call Guard** — `/live`, `/api/live-analyze` 실시간 청크 검출\n\n통합 가이드: [`docs/INTEGRATION_GUIDE.md`](https://github.com/example/scamguardian-v2/blob/main/docs/INTEGRATION_GUIDE.md)",
    "title": "ScamGuardian Signal Detection API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/admin/runs": {
      "get": {
        "description": "라벨링 대기/진행/완료 큐 목록. `claimed_by` 로 누가 작업 중인지 표시.\n\n**Query**: `status` (in_progress|done|None), `limit` (기본 50), `offset`.",
        "operationId": "admin_list_runs_api_admin_runs_get",
        "parameters": [
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Admin List Runs Api Admin Runs Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "description": "DB 미설정 (SCAMGUARDIAN_DATABASE_URL 또는 SQLITE_PATH)"
          },
          "401": {
            "description": "어드민 토큰 누락 또는 무효"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "500": {
            "description": "서버 내부 오류"
          }
        },
        "summary": "라벨링 큐 목록",
        "tags": [
          "Admin — Labeling"
        ]
      }
    },
    "/api/admin/runs/next": {
      "get": {
        "description": "아직 사람 annotation 이 없는 가장 오래된 run 1개. `options` 필드는 분류 라벨/플래그 카탈로그.",
        "operationId": "admin_next_run_api_admin_runs_next_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Admin Next Run Api Admin Runs Next Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "description": "DB 미설정 (SCAMGUARDIAN_DATABASE_URL 또는 SQLITE_PATH)"
          },
          "401": {
            "description": "어드민 토큰 누락 또는 무효"
          },
          "500": {
            "description": "서버 내부 오류"
          }
        },
        "summary": "다음 미라벨 run 자동 할당",
        "tags": [
          "Admin — Labeling"
        ]
      }
    },
    "/api/admin/runs/search": {
      "get": {
        "description": "transcript / scam_type / 라벨 여부로 필터링. `labeled=true` 면 사람 라벨 있는 것만, `false` 면 미라벨만.",
        "operationId": "admin_search_runs_api_admin_runs_search_get",
        "parameters": [
          {
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Q"
            }
          },
          {
            "in": "query",
            "name": "scam_type",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Scam Type"
            }
          },
          {
            "in": "query",
            "name": "labeled",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Labeled"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 30,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Admin Search Runs Api Admin Runs Search Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "description": "DB 미설정 (SCAMGUARDIAN_DATABASE_URL 또는 SQLITE_PATH)"
          },
          "401": {
            "description": "어드민 토큰 누락 또는 무효"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "500": {
            "description": "서버 내부 오류"
          }
        },
        "summary": "run 검색 (필터 다중)",
        "tags": [
          "Admin — Labeling"
        ]
      }
    },
    "/api/admin/runs/{run_id}": {
      "get": {
        "description": "분석 run + 사람 annotation + 사용자 챗봇 컨텍스트 전체. `metadata.user_context.qa_pairs` (사용자 Q&A) + `metadata.chat_history` (전체 대화) 노출.",
        "operationId": "admin_run_detail_api_admin_runs__run_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Admin Run Detail Api Admin Runs  Run Id  Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "description": "DB 미설정 (SCAMGUARDIAN_DATABASE_URL 또는 SQLITE_PATH)"
          },
          "401": {
            "description": "어드민 토큰 누락 또는 무효"
          },
          "404": {
            "description": "run not found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "500": {
            "description": "서버 내부 오류"
          }
        },
        "summary": "run 상세 (라벨링 에디터용)",
        "tags": [
          "Admin — Labeling"
        ]
      }
    },
    "/api/admin/runs/{run_id}/ai-draft": {
      "post": {
        "description": "Claude API 로 사람 라벨 *초안* 을 생성한다. 검수자는 초안 확인 후 수정/승인만. 응답 `draft` 는 `entities/triggered_flags` 에 `source: \"ai-draft\"` 태깅됨.",
        "operationId": "admin_ai_draft_api_admin_runs__run_id__ai_draft_post",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Admin Ai Draft Api Admin Runs  Run Id  Ai Draft Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "description": "DB 미설정 (SCAMGUARDIAN_DATABASE_URL 또는 SQLITE_PATH)"
          },
          "401": {
            "description": "어드민 토큰 누락 또는 무효"
          },
          "404": {
            "description": "run not found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "500": {
            "description": "서버 내부 오류"
          }
        },
        "summary": "Claude 라벨링 초안 생성",
        "tags": [
          "Admin — Labeling"
        ]
      }
    },
    "/api/admin/runs/{run_id}/annotations": {
      "post": {
        "description": "검수자의 정답 라벨을 저장. 기존 annotation 있으면 덮어쓰기.\n\n**Body** (`HumanAnnotationRequest`): content_label, scam_type_gt(scam_attempt 일 때만 필수), entities_gt, triggered_flags_gt, transcript_corrected_text(선택), stt_quality(1~5), notes, sample_kind, source_ref. content_label 미지정 시 학습 로더가 fallback 적용.",
        "operationId": "admin_save_annotation_api_admin_runs__run_id__annotations_post",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HumanAnnotationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Admin Save Annotation Api Admin Runs  Run Id  Annotations Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "description": "DB 미설정 (SCAMGUARDIAN_DATABASE_URL 또는 SQLITE_PATH)"
          },
          "401": {
            "description": "어드민 토큰 누락 또는 무효"
          },
          "404": {
            "description": "run not found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "500": {
            "description": "서버 내부 오류"
          }
        },
        "summary": "정답 라벨 저장 (upsert)",
        "tags": [
          "Admin — Labeling"
        ]
      }
    },
    "/api/admin/runs/{run_id}/claim": {
      "post": {
        "description": "검수자 이름으로 run 을 점유한다. `claimed_by` / `claimed_at` 컬럼 갱신. TTL 30분 — 다른 검수자가 해당 시간 안에 다시 claim 할 수 없다.\n\n**Body**: `{labeler: str}` — 빈 문자열이면 `\"Admin\"` 으로 저장.\n\n**409**: 다른 검수자가 이미 점유 중.",
        "operationId": "admin_claim_run_api_admin_runs__run_id__claim_post",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ClaimRunRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Admin Claim Run Api Admin Runs  Run Id  Claim Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "description": "DB 미설정 (SCAMGUARDIAN_DATABASE_URL 또는 SQLITE_PATH)"
          },
          "401": {
            "description": "어드민 토큰 누락 또는 무효"
          },
          "409": {
            "description": "다른 검수자가 점유 중"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "500": {
            "description": "서버 내부 오류"
          }
        },
        "summary": "run 검수 claim",
        "tags": [
          "Admin — Labeling"
        ]
      }
    },
    "/api/admin/runs/{run_id}/media": {
      "get": {
        "description": "라벨링 도중 STT 정확도 검증용으로 원본 음성/영상/이미지 파일을 그대로 반환. path traversal 방지 — `.scamguardian/uploads/` 안 파일만 허용.",
        "operationId": "admin_get_media_api_admin_runs__run_id__media_get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "description": "허용되지 않은 경로 (path traversal)"
          },
          "401": {
            "description": "어드민 토큰 누락 또는 무효"
          },
          "404": {
            "description": "media 파일 없음"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "500": {
            "description": "서버 내부 오류"
          }
        },
        "summary": "원본 업로드 파일 스트리밍",
        "tags": [
          "Admin — Labeling"
        ]
      }
    },
    "/api/admin/scam-types": {
      "get": {
        "description": "DEFAULT_SCAM_TYPES 12종 외 어드민이 추가한 커스텀 분류 카탈로그.",
        "operationId": "admin_scam_types_api_admin_scam_types_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Admin Scam Types Api Admin Scam Types Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "description": "DB 미설정 (SCAMGUARDIAN_DATABASE_URL 또는 SQLITE_PATH)"
          },
          "401": {
            "description": "어드민 토큰 누락 또는 무효"
          },
          "500": {
            "description": "서버 내부 오류"
          }
        },
        "summary": "사용자 정의 스캠 유형 목록",
        "tags": [
          "Admin — Labeling"
        ]
      },
      "post": {
        "description": "런타임 분류 카탈로그에 새 scam_type 추가. 기본 12종과 같은 이름은 거절(400). 추가 즉시 `get_runtime_scam_taxonomy()` 에 반영되어 분류기·라벨링 UI 가 인식.",
        "operationId": "admin_add_scam_type_api_admin_scam_types_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScamTypeCatalogRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Admin Add Scam Type Api Admin Scam Types Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "description": "이름 중복 또는 유효성 실패"
          },
          "401": {
            "description": "어드민 토큰 누락 또는 무효"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "500": {
            "description": "서버 내부 오류"
          }
        },
        "summary": "사용자 정의 스캠 유형 추가",
        "tags": [
          "Admin — Labeling"
        ]
      }
    },
    "/api/admin/stats": {
      "get": {
        "description": "총 run 수 / 라벨 진행률 / scam_type 분포 등 대시보드 카운터.",
        "operationId": "admin_stats_api_admin_stats_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Admin Stats Api Admin Stats Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "description": "DB 미설정 (SCAMGUARDIAN_DATABASE_URL 또는 SQLITE_PATH)"
          },
          "401": {
            "description": "어드민 토큰 누락 또는 무효"
          },
          "500": {
            "description": "서버 내부 오류"
          }
        },
        "summary": "대시보드 통계",
        "tags": [
          "Admin — Labeling"
        ]
      }
    },
...
  },
...
}